requirements.txt
Flask
Flask-CORS
Pillow
rembg
onnxruntime
gunicorn